requirements.txt
streamlit
numpy
matplotlib
paho-mqtt
scipy